backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
pydantic-settings==2.7.1
composio-openai==0.7.3
openai==1.59.9
httpx==0.28.1
python-dotenv==1.0.1

